requirements.txt
asgiref==3.7.2
Django==3.2.23
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0
pillow
django_next_prev
django-extensions
django-robots
django-debug-toolbar
django-taggit